+++ cmd/dendrite-polylith-multi/personalities/federationsender.go
@@ -1,4 +1,4 @@
-// Copyright 2017 Vector Creations Ltd
+// Copyright 2020 The Matrix.org Foundation C.I.C.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,33 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package personalities
 
 import (
+	"github.com/matrix-org/dendrite/federationsender"
+	"github.com/matrix-org/dendrite/internal/config"
 	"github.com/matrix-org/dendrite/internal/setup"
-	"github.com/matrix-org/dendrite/syncapi"
 )
 
-func main() {
-	cfg := setup.ParseFlags(false)
-
-	base := setup.NewBaseDendrite(cfg, "SyncAPI", true)
-	defer base.Close() // nolint: errcheck
-
-	userAPI := base.UserAPIClient()
+func FederationSender(base *setup.BaseDendrite, cfg *config.Dendrite) {
 	federation := base.CreateFederationClient()
 
-	rsAPI := base.RoomserverHTTPClient()
+	serverKeyAPI := base.SigningKeyServerHTTPClient()
+	keyRing := serverKeyAPI.KeyRing()
 
-	syncapi.AddPublicRoutes(
-		base.PublicClientAPIMux, userAPI, rsAPI,
-		base.KeyServerHTTPClient(),
-		federation, &cfg.SyncAPI,
+	rsAPI := base.RoomserverHTTPClient()
+	fsAPI := federationsender.NewInternalAPI(
+		base, federation, rsAPI, keyRing,
 	)
+	federationsender.AddInternalRoutes(base.InternalAPIMux, fsAPI)
 
 	base.SetupAndServeHTTP(
-		base.Cfg.SyncAPI.InternalAPI.Listen,
-		base.Cfg.SyncAPI.ExternalAPI.Listen,
+		base.Cfg.FederationSender.InternalAPI.Listen, // internal listener
+		setup.NoListener, // external listener
 		nil, nil,
 	)
 }
